Roughly refactored codebase for easier readability and maintaining
