+++ .env.sample
@@ -6,4 +6,3 @@ TM_SERVER_ACC_PASSWORD=
 TM_OAUTH2_CLIENT_ID=
 TM_OAUTH2_CLIENT_SECRET=
 PORT=
-MONGODB_URI=
