+++ package.json
@@ -15,10 +15,10 @@
         "discord-interactions": "^3.4.0",
         "dotenv": "^16.3.1",
         "express": "^4.18.2",
-        "mongodb": "^6.10.0",
         "node-cache": "^5.1.2",
         "node-fetch": "^3.3.2",
         "node-schedule": "^2.1.1",
+        "sqlite3": "^5.1.7",
         "winston": "^3.17.0"
     },
     "devDependencies": {
